Finished
Can now edit submissions on forms that should only accepts a submission once and allow the editing of said submission.

You can submit multiple times in a row, won't allow yo uto go back to submit page if you aren't allowed to submit multiple times!

Extra addition will be rating handling with extra handles for copying over ratings to the databse table.

Will also soon add useful methods for using forms in the backend. Otherwise frontend is done. (It's missing some features, but these can be toggled in the admin view and are just extra features that are avilable but not needed for the prototype!)